Add osv output lockfile
