+++ pkg/lockfile/fixtures/osvscannerresults/empty.json
@@ -1,0 +1,3 @@
+{
+    "results": []
+}
