+++ pkg/lockfile/fixtures/osvscannerresults/not-json.txt
@@ -1,0 +1,1 @@
+this is not valid json! (I think)
