+++ pkg/lockfile/parse-requirements-txt_test.go
@@ -106,6 +106,7 @@ func TestParseRequirementsTxt_OneRequirementUnconstrained(t *testing.T) {
 			Version:   "0.0.0",
 			Ecosystem: lockfile.PipEcosystem,
 			CompareAs: lockfile.PipEcosystem,
+			Source:    "%%/fixtures/pip/one-package-unconstrained.txt",
 		},
 	})
 }
@@ -125,12 +126,13 @@ func TestParseRequirementsTxt_OneRequirementConstrained(t *testing.T) {
 			Version:   "2.2.24",
 			Ecosystem: lockfile.PipEcosystem,
 			CompareAs: lockfile.PipEcosystem,
+			Source:    "%%/fixtures/pip/one-package-constrained.txt",
 		},
 	})
 }
 
 func TestParseRequirementsTxt_MultipleRequirementsConstrained(t *testing.T) {
-	t.Parallel()
+	// t.Parallel()
 
 	packages, err := lockfile.ParseRequirementsTxt("fixtures/pip/multiple-packages-constrained.txt")
 
@@ -144,78 +146,91 @@ func TestParseRequirementsTxt_MultipleRequirementsConstrained(t *testing.T) {
 			Version:   "2.5.1",
 			Ecosystem: lockfile.PipEcosystem,
 			CompareAs: lockfile.PipEcosystem,
+			Source:    "%%/fixtures/pip/multiple-packages-constrained.txt",
 		},
 		{
 			Name:      "beautifulsoup4",
 			Version:   "4.9.3",
 			Ecosystem: lockfile.PipEcosystem,
 			CompareAs: lockfile.PipEcosystem,
+			Source:    "%%/fixtures/pip/multiple-packages-constrained.txt",
 		},
 		{
 			Name:      "boto3",
 			Version:   "1.17.19",
 			Ecosystem: lockfile.PipEcosystem,
 			CompareAs: lockfile.PipEcosystem,
+			Source:    "%%/fixtures/pip/multiple-packages-constrained.txt",
 		},
 		{
 			Name:      "botocore",
 			Version:   "1.20.19",
 			Ecosystem: lockfile.PipEcosystem,
 			CompareAs: lockfile.PipEcosystem,
+			Source:    "%%/fixtures/pip/multiple-packages-constrained.txt",
 		},
 		{
 			Name:      "certifi",
 			Version:   "2020.12.5",
 			Ecosystem: lockfile.PipEcosystem,
 			CompareAs: lockfile.PipEcosystem,
+			Source:    "%%/fixtures/pip/multiple-packages-constrained.txt",
 		},
 		{
 			Name:      "chardet",
 			Version:   "4.0.0",
 			Ecosystem: lockfile.PipEcosystem,
 			CompareAs: lockfile.PipEcosystem,
+			Source:    "%%/fixtures/pip/multiple-packages-constrained.txt",
 		},
 		{
 			Name:      "circus",
 			Version:   "0.17.1",
 			Ecosystem: lockfile.PipEcosystem,
 			CompareAs: lockfile.PipEcosystem,
+			Source:    "%%/fixtures/pip/multiple-packages-constrained.txt",
 		},
 		{
 			Name:      "click",
 			Version:   "7.1.2",
 			Ecosystem: lockfile.PipEcosystem,
 			CompareAs: lockfile.PipEcosystem,
+			Source:    "%%/fixtures/pip/multiple-packages-constrained.txt",
 		},
 		{
 			Name:      "django-debug-toolbar",
 			Version:   "3.2.1",
 			Ecosystem: lockfile.PipEcosystem,
 			CompareAs: lockfile.PipEcosystem,
+			Source:    "%%/fixtures/pip/multiple-packages-constrained.txt",
 		},
 		{
 			Name:      "django-filter",
 			Version:   "2.4.0",
 			Ecosystem: lockfile.PipEcosystem,
 			CompareAs: lockfile.PipEcosystem,
+			Source:    "%%/fixtures/pip/multiple-packages-constrained.txt",
 		},
 		{
 			Name:      "django-nose",
 			Version:   "1.4.7",
 			Ecosystem: lockfile.PipEcosystem,
 			CompareAs: lockfile.PipEcosystem,
+			Source:    "%%/fixtures/pip/multiple-packages-constrained.txt",
 		},
 		{
 			Name:      "django-storages",
 			Version:   "1.11.1",
 			Ecosystem: lockfile.PipEcosystem,
 			CompareAs: lockfile.PipEcosystem,
+			Source:    "%%/fixtures/pip/multiple-packages-constrained.txt",
 		},
 		{
 			Name:      "django",
 			Version:   "2.2.24",
 			Ecosystem: lockfile.PipEcosystem,
 			CompareAs: lockfile.PipEcosystem,
+			Source:    "%%/fixtures/pip/multiple-packages-constrained.txt",
 		},
 	})
 }
@@ -235,48 +250,56 @@ func TestParseRequirementsTxt_MultipleRequirementsMixed(t *testing.T) {
 			Version:   "0.0.0",
 			Ecosystem: lockfile.PipEcosystem,
 			CompareAs: lockfile.PipEcosystem,
+			Source:    "%%/fixtures/pip/multiple-packages-mixed.txt",
 		},
 		{
 			Name:      "flask-cors",
 			Version:   "0.0.0",
 			Ecosystem: lockfile.PipEcosystem,
 			CompareAs: lockfile.PipEcosystem,
+			Source:    "%%/fixtures/pip/multiple-packages-mixed.txt",
 		},
 		{
 			Name:      "pandas",
 			Version:   "0.23.4",
 			Ecosystem: lockfile.PipEcosystem,
 			CompareAs: lockfile.PipEcosystem,
+			Source:    "%%/fixtures/pip/multiple-packages-mixed.txt",
 		},
 		{
 			Name:      "numpy",
 			Version:   "1.16.0",
 			Ecosystem: lockfile.PipEcosystem,
 			CompareAs: lockfile.PipEcosystem,
+			Source:    "%%/fixtures/pip/multiple-packages-mixed.txt",
 		},
 		{
 			Name:      "scikit-learn",
 			Version:   "0.20.1",
 			Ecosystem: lockfile.PipEcosystem,
 			CompareAs: lockfile.PipEcosystem,
+			Source:    "%%/fixtures/pip/multiple-packages-mixed.txt",
 		},
 		{
 			Name:      "sklearn",
 			Version:   "0.0.0",
 			Ecosystem: lockfile.PipEcosystem,
 			CompareAs: lockfile.PipEcosystem,
+			Source:    "%%/fixtures/pip/multiple-packages-mixed.txt",
 		},
 		{
 			Name:      "requests",
 			Version:   "0.0.0",
 			Ecosystem: lockfile.PipEcosystem,
 			CompareAs: lockfile.PipEcosystem,
+			Source:    "%%/fixtures/pip/multiple-packages-mixed.txt",
 		},
 		{
 			Name:      "gevent",
 			Version:   "0.0.0",
 			Ecosystem: lockfile.PipEcosystem,
 			CompareAs: lockfile.PipEcosystem,
+			Source:    "%%/fixtures/pip/multiple-packages-mixed.txt",
 		},
 	})
 }
@@ -296,60 +319,70 @@ func TestParseRequirementsTxt_FileFormatExample(t *testing.T) {
 			Version:   "0.0.0",
 			Ecosystem: lockfile.PipEcosystem,
 			CompareAs: lockfile.PipEcosystem,
+			Source:    "%%/fixtures/pip/file-format-example.txt",
 		},
 		{
 			Name:      "pytest-cov",
 			Version:   "0.0.0",
 			Ecosystem: lockfile.PipEcosystem,
 			CompareAs: lockfile.PipEcosystem,
+			Source:    "%%/fixtures/pip/file-format-example.txt",
 		},
 		{
 			Name:      "beautifulsoup4",
 			Version:   "0.0.0",
 			Ecosystem: lockfile.PipEcosystem,
 			CompareAs: lockfile.PipEcosystem,
+			Source:    "%%/fixtures/pip/file-format-example.txt",
 		},
 		{
 			Name:      "docopt",
 			Version:   "0.6.1",
 			Ecosystem: lockfile.PipEcosystem,
 			CompareAs: lockfile.PipEcosystem,
+			Source:    "%%/fixtures/pip/file-format-example.txt",
 		},
 		{
 			Name:      "keyring",
 			Version:   "4.1.1",
 			Ecosystem: lockfile.PipEcosystem,
 			CompareAs: lockfile.PipEcosystem,
+			Source:    "%%/fixtures/pip/file-format-example.txt",
 		},
 		{
 			Name:      "coverage",
 			Version:   "0.0.0",
 			Ecosystem: lockfile.PipEcosystem,
 			CompareAs: lockfile.PipEcosystem,
+			Source:    "%%/fixtures/pip/file-format-example.txt",
 		},
 		{
 			Name:      "mopidy-dirble",
 			Version:   "1.1",
 			Ecosystem: lockfile.PipEcosystem,
 			CompareAs: lockfile.PipEcosystem,
+			Source:    "%%/fixtures/pip/file-format-example.txt",
 		},
 		{
 			Name:      "rejected",
 			Version:   "0.0.0",
 			Ecosystem: lockfile.PipEcosystem,
 			CompareAs: lockfile.PipEcosystem,
+			Source:    "%%/fixtures/pip/file-format-example.txt",
 		},
 		{
 			Name:      "green",
 			Version:   "0.0.0",
 			Ecosystem: lockfile.PipEcosystem,
 			CompareAs: lockfile.PipEcosystem,
+			Source:    "%%/fixtures/pip/file-format-example.txt",
 		},
 		{
 			Name:      "django",
 			Version:   "2.2.24",
 			Ecosystem: lockfile.PipEcosystem,
 			CompareAs: lockfile.PipEcosystem,
+			Source:    "%%/fixtures/pip/other-file.txt",
 		},
 	})
 }
@@ -369,6 +402,7 @@ func TestParseRequirementsTxt_WithAddedSupport(t *testing.T) {
 			Version:   "20.3.0",
 			Ecosystem: lockfile.PipEcosystem,
 			CompareAs: lockfile.PipEcosystem,
+			Source:    "%%/fixtures/pip/with-added-support.txt",
 		},
 	})
 }
@@ -388,18 +422,21 @@ func TestParseRequirementsTxt_NonNormalizedNames(t *testing.T) {
 			Version:   "5.4.0",
 			Ecosystem: lockfile.PipEcosystem,
 			CompareAs: lockfile.PipEcosystem,
+			Source:    "%%/fixtures/pip/non-normalized-names.txt",
 		},
 		{
 			Name:      "pillow",
 			Version:   "1.0.0",
 			Ecosystem: lockfile.PipEcosystem,
 			CompareAs: lockfile.PipEcosystem,
+			Source:    "%%/fixtures/pip/non-normalized-names.txt",
 		},
 		{
 			Name:      "twisted",
 			Version:   "20.3.0",
 			Ecosystem: lockfile.PipEcosystem,
 			CompareAs: lockfile.PipEcosystem,
+			Source:    "%%/fixtures/pip/non-normalized-names.txt",
 		},
 	})
 }
@@ -419,60 +456,70 @@ func TestParseRequirementsTxt_WithMultipleROptions(t *testing.T) {
 			Version:   "0.0.0",
 			Ecosystem: lockfile.PipEcosystem,
 			CompareAs: lockfile.PipEcosystem,
+			Source:    "%%/fixtures/pip/multiple-packages-mixed.txt",
 		},
 		{
 			Name:      "flask-cors",
 			Version:   "0.0.0",
 			Ecosystem: lockfile.PipEcosystem,
 			CompareAs: lockfile.PipEcosystem,
+			Source:    "%%/fixtures/pip/multiple-packages-mixed.txt",
 		},
 		{
 			Name:      "pandas",
 			Version:   "0.23.4",
 			Ecosystem: lockfile.PipEcosystem,
 			CompareAs: lockfile.PipEcosystem,
+			Source:    "%%/fixtures/pip/with-multiple-r-options.txt",
 		},
 		{
 			Name:      "numpy",
 			Version:   "1.16.0",
 			Ecosystem: lockfile.PipEcosystem,
 			CompareAs: lockfile.PipEcosystem,
+			Source:    "%%/fixtures/pip/multiple-packages-mixed.txt",
 		},
 		{
 			Name:      "scikit-learn",
 			Version:   "0.20.1",
 			Ecosystem: lockfile.PipEcosystem,
 			CompareAs: lockfile.PipEcosystem,
+			Source:    "%%/fixtures/pip/multiple-packages-mixed.txt",
 		},
 		{
 			Name:      "sklearn",
 			Version:   "0.0.0",
 			Ecosystem: lockfile.PipEcosystem,
 			CompareAs: lockfile.PipEcosystem,
+			Source:    "%%/fixtures/pip/multiple-packages-mixed.txt",
 		},
 		{
 			Name:      "requests",
 			Version:   "0.0.0",
 			Ecosystem: lockfile.PipEcosystem,
 			CompareAs: lockfile.PipEcosystem,
+			Source:    "%%/fixtures/pip/multiple-packages-mixed.txt",
 		},
 		{
 			Name:      "gevent",
 			Version:   "0.0.0",
 			Ecosystem: lockfile.PipEcosystem,
 			CompareAs: lockfile.PipEcosystem,
+			Source:    "%%/fixtures/pip/multiple-packages-mixed.txt",
 		},
 		{
 			Name:      "requests",
 			Version:   "1.2.3",
 			Ecosystem: lockfile.PipEcosystem,
 			CompareAs: lockfile.PipEcosystem,
+			Source:    "%%/fixtures/pip/with-multiple-r-options.txt",
 		},
 		{
 			Name:      "django",
 			Version:   "2.2.24",
 			Ecosystem: lockfile.PipEcosystem,
 			CompareAs: lockfile.PipEcosystem,
+			Source:    "%%/fixtures/pip/one-package-constrained.txt",
 		},
 	})
 }
@@ -501,24 +548,28 @@ func TestParseRequirementsTxt_DuplicateROptions(t *testing.T) {
 			Version:   "0.1.0",
 			Ecosystem: lockfile.PipEcosystem,
 			CompareAs: lockfile.PipEcosystem,
+			Source:    "%%/fixtures/pip/duplicate-r-base.txt",
 		},
 		{
 			Name:      "pandas",
 			Version:   "0.23.4",
 			Ecosystem: lockfile.PipEcosystem,
 			CompareAs: lockfile.PipEcosystem,
+			Source:    "%%/fixtures/pip/duplicate-r-dev.txt",
 		},
 		{
 			Name:      "requests",
 			Version:   "1.2.3",
 			Ecosystem: lockfile.PipEcosystem,
 			CompareAs: lockfile.PipEcosystem,
+			Source:    "%%/fixtures/pip/duplicate-r-dev.txt",
 		},
 		{
 			Name:      "unittest",
 			Version:   "1.0.0",
 			Ecosystem: lockfile.PipEcosystem,
 			CompareAs: lockfile.PipEcosystem,
+			Source:    "%%/fixtures/pip/duplicate-r-test.txt",
 		},
 	})
 }
@@ -538,12 +589,14 @@ func TestParseRequirementsTxt_CyclicRSelf(t *testing.T) {
 			Version:   "0.23.4",
 			Ecosystem: lockfile.PipEcosystem,
 			CompareAs: lockfile.PipEcosystem,
+			Source:    "%%/fixtures/pip/cyclic-r-self.txt",
 		},
 		{
 			Name:      "requests",
 			Version:   "1.2.3",
 			Ecosystem: lockfile.PipEcosystem,
 			CompareAs: lockfile.PipEcosystem,
+			Source:    "%%/fixtures/pip/cyclic-r-self.txt",
 		},
 	})
 }
@@ -563,18 +616,21 @@ func TestParseRequirementsTxt_CyclicRComplex(t *testing.T) {
 			Version:   "1",
 			Ecosystem: lockfile.PipEcosystem,
 			CompareAs: lockfile.PipEcosystem,
+			Source:    "%%/fixtures/pip/cyclic-r-complex-1.txt",
 		},
 		{
 			Name:      "cyclic-r-complex",
 			Version:   "2",
 			Ecosystem: lockfile.PipEcosystem,
 			CompareAs: lockfile.PipEcosystem,
+			Source:    "%%/fixtures/pip/cyclic-r-complex-2.txt",
 		},
 		{
 			Name:      "cyclic-r-complex",
 			Version:   "3",
 			Ecosystem: lockfile.PipEcosystem,
 			CompareAs: lockfile.PipEcosystem,
+			Source:    "%%/fixtures/pip/cyclic-r-complex-3.txt",
 		},
 	})
 }
@@ -594,24 +650,28 @@ func TestParseRequirementsTxt_WithPerRequirementOptions(t *testing.T) {
 			Version:   "1.26.121",
 			Ecosystem: lockfile.PipEcosystem,
 			CompareAs: lockfile.PipEcosystem,
+			Source:    "%%/fixtures/pip/with-per-requirement-options.txt",
 		},
 		{
 			Name:      "foo",
 			Version:   "1.0.0",
 			Ecosystem: lockfile.PipEcosystem,
 			CompareAs: lockfile.PipEcosystem,
+			Source:    "%%/fixtures/pip/with-per-requirement-options.txt",
 		},
 		{
 			Name:      "fooproject",
 			Version:   "1.2",
 			Ecosystem: lockfile.PipEcosystem,
 			CompareAs: lockfile.PipEcosystem,
+			Source:    "%%/fixtures/pip/with-per-requirement-options.txt",
 		},
 		{
 			Name:      "barproject",
 			Version:   "1.2",
 			Ecosystem: lockfile.PipEcosystem,
 			CompareAs: lockfile.PipEcosystem,
+			Source:    "%%/fixtures/pip/with-per-requirement-options.txt",
 		},
 	})
 }
@@ -631,24 +691,28 @@ func TestParseRequirementsTxt_LineContinuation(t *testing.T) {
 			Version:   "1.2.3",
 			Ecosystem: lockfile.PipEcosystem,
 			CompareAs: lockfile.PipEcosystem,
+			Source:    "%%/fixtures/pip/line-continuation.txt",
 		},
 		{
 			Name:      "bar",
 			Version:   "4.5\\\\",
 			Ecosystem: lockfile.PipEcosystem,
 			CompareAs: lockfile.PipEcosystem,
+			Source:    "%%/fixtures/pip/line-continuation.txt",
 		},
 		{
 			Name:      "baz",
 			Version:   "7.8.9",
 			Ecosystem: lockfile.PipEcosystem,
 			CompareAs: lockfile.PipEcosystem,
+			Source:    "%%/fixtures/pip/line-continuation.txt",
 		},
 		{
 			Name:      "qux",
 			Version:   "10.11.12",
 			Ecosystem: lockfile.PipEcosystem,
 			CompareAs: lockfile.PipEcosystem,
+			Source:    "%%/fixtures/pip/line-continuation.txt",
 		},
 	})
 }
