+++ pkg/models/vulnerability.go
@@ -2,14 +2,7 @@ package models
 
 import (
 	"encoding/json"
-	"fmt"
-	"os"
-	"sort"
 	"time"
-
-	"github.com/google/osv-scanner/internal/semantic"
-	"github.com/google/osv-scanner/pkg/lockfile"
-	"golang.org/x/exp/slices"
 )
 
 // Package identifies the affected code library or command provided by the
@@ -173,144 +166,4 @@ func (v Vulnerability) MarshalYAML() (interface{}, error) {
 	}
 
 	return raw, nil
-}
-
-func (e Event) version() string {
-	if e.Introduced != "" {
-		return e.Introduced
-	}
-
-	if e.Fixed != "" {
-		return e.Fixed
-	}
-
-	if e.Limit != "" {
-		return e.Limit
-	}
-
-	if e.LastAffected != "" {
-		return e.LastAffected
-	}
-
-	return ""
-}
-
-func (ar Range) containsVersion(pkg lockfile.PackageDetails) bool {
-	if ar.Type != RangeEcosystem && ar.Type != RangeSemVer {
-		return false
-	}
-	// todo: we should probably warn here
-	if len(ar.Events) == 0 {
-		return false
-	}
-
-	vp := semantic.MustParse(pkg.Version, pkg.CompareAs)
-
-	sort.Slice(ar.Events, func(i, j int) bool {
-		a := ar.Events[i]
-		b := ar.Events[j]
-
-		if a.Introduced == "0" {
-			return true
-		}
-
-		if b.Introduced == "0" {
-			return false
-		}
-
-		return semantic.MustParse(a.version(), pkg.CompareAs).CompareStr(b.version()) < 0
-	})
-
-	var affected bool
-	for _, e := range ar.Events {
-		if affected {
-			if e.Fixed != "" {
-				affected = vp.CompareStr(e.Fixed) < 0
-			} else if e.LastAffected != "" {
-				affected = e.LastAffected == pkg.Version || vp.CompareStr(e.LastAffected) <= 0
-			}
-		} else if e.Introduced != "" {
-			affected = e.Introduced == "0" || vp.CompareStr(e.Introduced) >= 0
-		}
-	}
-
-	return affected
-}
-
-// affectsVersion checks if the given version is within the range
-// specified by the events of any "Ecosystem" or "Semver" type ranges
-func affectsVersion(a []Range, pkg lockfile.PackageDetails) bool {
-	for _, r := range a {
-		if r.Type != RangeEcosystem && r.Type != RangeSemVer {
-			return false
-		}
-		if r.containsVersion(pkg) {
-			return true
-		}
-	}
-
-	return false
-}
-
-func (v Vulnerability) isAliasOfID(id string) bool {
-	for _, alias := range v.Aliases {
-		if alias == id {
-			return true
-		}
-	}
-
-	return false
-}
-
-func (v Vulnerability) isAliasOf(vulnerability Vulnerability) bool {
-	for _, alias := range vulnerability.Aliases {
-		if v.ID == alias || v.isAliasOfID(alias) {
-			return true
-		}
-	}
-
-	return false
-}
-
-func (v Vulnerability) AffectsEcosystem(ecosystem lockfile.Ecosystem) bool {
-	for _, affected := range v.Affected {
-		if string(affected.Package.Ecosystem) == string(ecosystem) {
-			return true
-		}
-	}
-
-	return false
-}
-
-func (v Vulnerability) IsAffected(pkg lockfile.PackageDetails) bool {
-	for _, affected := range v.Affected {
-		if string(affected.Package.Ecosystem) == string(pkg.Ecosystem) &&
-			affected.Package.Name == pkg.Name {
-			if len(affected.Ranges) == 0 && len(affected.Versions) == 0 {
-				_, _ = fmt.Fprintf(
-					os.Stderr,
-					"%s does not have any ranges or versions - this is probably a mistake!\n",
-					v.ID,
-				)
-
-				continue
-			}
-
-			if slices.Contains(affected.Versions, pkg.Version) {
-				return true
-			}
-
-			if affectsVersion(affected.Ranges, pkg) {
-				return true
-			}
-
-			// if a package does not have a version, assume it is vulnerable
-			// as false positives are better than false negatives here
-			if pkg.Version == "" {
-				return true
-			}
-		}
-	}
-
-	return false
 }
